requirements.txt
streamlit>=1.31
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
langgraph>=0.1
qdrant-client>=1.7
python-dotenv>=1.0
pydantic>=2.6
